refactor(player-profile): lift the last three logic clusters out
Takes `PlayerProfile.tsx` 636 → 460, and to zero `useState` and zero `invoke`
calls — everything it renders now comes from a hook or a prop.

**`usePlayerProfileData`** — the two things a profile fetches that the game
state does not already hold: the backend's advanced-stats view and the player's
last few matches. Both render from squad data first and upgrade in place, and
both keep the previous object identity when the new payload is equivalent,
which the extraction makes visible rather than incidental. The
`areAdvancedStatsEqual` comparison moves with them, and the array comparison
becomes a named `haveSameFixtures` instead of an inline `.every`.

**`useScoutPlayerFlow`** — sending a scout was a ~20-line async IIFE inside a
JSX prop. `scoutAvailability` travels with it because it is derived from the
scout status as well as the squad: a scout already on their way is not
available to send again.

**`useInitialModalIntent`** — the renewal and termination auto-open effects were
structurally identical, down to the four-clause guard and the consumed flag that
stops a modal reopening every time the player closes it. Now one parameterised
hook, used twice.

The reset key is deliberately shared between the two calls rather than split per
modal. Splitting it reads more correct, but it would change behaviour: today,
changing either prop re-arms both intents, because one effect reset both flags.
That coupling is preserved here and left for a PR that can be about it.

Covered by the three auto-open tests added earlier in this branch, which exist
precisely because nothing reached those effects before.
